dspinner: RecursiveKeys(): do not hang on cancellations
Per https://github.com/ipfs/kubo/issues/10593, if no one
is reading from the channel returned by RecursiveKeys() and the context is cancelled, streamIndex will hang indefinitely.

Proposed fix is to always select when attempting to write to the `out` channel. If the context is done and there is no one to read, we can abort.